# Conflicts:
#	trans_server/transaction_server.py

+++ trans_server/transaction_server.py
@@ -161,7 +161,7 @@ class TransactionServer:
         succeeded = False
         user = data["userid"]
         symbol = data["StockSymbol"]
-        amount = int(data["amount"])
+        amount = float(data["amount"])
         cli_data = self.cli_data
 
         if cli_data.rem_stock(user, symbol, amount):
